New notebook nearestneighbors
This notebook builds on the discovery in jaccard notebook and focuses
on nearest neighbor search

Initial version includes k=100 NN search using full playlist based off
work from jaccard notebook.
It extends the search using a sample of the playlist with 5 entries. The
current result set returns on 5 neighbors instead of the requested 100.